refactor: move initColorScheme into theme module
